Check that all jobs scheduled are included as part of a system's dependencies
This relies on using OnUpdateForJob(), which takes the input JobHandle and returns a JobHandle.

The next step is to get rid of the AddDependency()/GetDependency() stuff and make it private, so there's only one way to accomplish the dependency graph construction.

Add validation that all jobs created in the above manner follow the rules.

Use new atomic safety handle APIs to spruce up error messages.