Add a rudimentary "new game" screen
I tried to do this with separate Worlds initially, so we could just
discard everything and start over, but it doesn't seem possible to do
this when the default world is also active and doing its thing.

If we had a custom world for each game session, we wouldn't need to
destroy all the enemies and shots when the player is destroyed, we'd
just discard the world.